requirements.txt
click==8.0.2
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.18.1
Werkzeug==2.0.2
